add: extra condition to check for browser or node environment
Popular JavaScript module bundlers like Webpack and Browserify expose a subset of native Node environment variables. This means that the condition typeof require === 'function' returns true when code is being processed in a module bundler. This means that code meant only for the Node environment is shipped to the browser.